+++ core/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. This projec
 
 ## 1.0.13 (in progress)
 
++ [#1855](https://github.com/luyadev/luya/issues/1855) If create a url to an other module, don't replace the url with current module context. 
+
 ## 1.0.12 (8. October 2018)
 
 + [#1856](https://github.com/luyadev/luya/issues/1856) If application console method is runing, the cli determination should be forced.